refactor(#181): 중복 로직 facade로 분리
